Refactoring Project Structure
Ofc so I've made changes like the way u want, Im also move translate option's array to separate folder to make it easy to audit in the future
