+++ src/components/base/Header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Link from '../Link';
+import Link from './Link';
 
 const Header = () => {
     return (
