+++ src/data/translateOptions.js
@@ -1,8 +1,4 @@
-import React, { useState } from "react";
-import Dropdown from "../Dropdown";
-import Convert from "../../base/Convert";
-
-const options = [
+export const translateOptions = [
     { 
         label:'Afrikaans',
         value:'af'
@@ -16,11 +12,11 @@ const options = [
         value: 'ar',
     },
     {
-	label: 'Bengali',
-	value: 'bn'
+        label: 'Bengali',
+        value: 'bn'
     },
     {
-	label: 'Bulgarian',
+	    label: 'Bulgarian',
         value: 'bg'
     },
     {
@@ -120,34 +116,3 @@ const options = [
         value: 'zu'
     },
 ]
-
-
-const Translate = () => {
-  const [language, setLanguage] = useState(options[0]);
-  const [text, setText] = useState("");
-
-  return (
-    <div>
-      <div className="ui form">
-        <div className="field">
-          <h4>
-            <label>Enter text</label>
-          </h4>
-          <input value={text} onChange={(e) => setText(e.target.value)} />
-        </div>
-      </div>
-      <Dropdown
-        options={options}
-        selected={language}
-        onSelectedChange={setLanguage}
-        label="Select a Language"
-        demo=""
-      />
-
-      <h3 className="ui header">Output</h3>
-      <Convert text={text} language={language} />
-    </div>
-  );
-};
-
-export default Translate;
